Add canonical window hierarchy deep links
